server/requirements.txt
flask==3.0.2
flask-cors==4.0.0
haystack-ai
elasticsearch-haystack
sentence-transformers
pdfplumber
fitz
PyMuPDF